Move utils and types to their own directories
