+++ src/anythingBut.ts
@@ -1,4 +1,4 @@
-import { simpleExp } from "./utils";
+import simpleExp from "./util/simple-exp";
 
 const anythingBut = simpleExp<string>(exp => `[^${exp}]*`);
 export default anythingBut;
