+++ src/oneOrMore.ts
@@ -1,4 +1,4 @@
-import { simpleExp } from "./utils";
+import simpleExp from "./util/simple-exp";
 
 const oneOrMore = simpleExp(exp => `${exp}+`);
 export default oneOrMore;
