+++ src/or.ts
@@ -1,5 +1,5 @@
-import { Expression } from "./types";
-import { simplifyAndGroup } from "./utils";
+import Expression from "./types/expression";
+import simplifyAndGroup from "./util/simplify-and-group";
 
 export default function or(...inputs: Expression[]) {
   return new RegExp(inputs.map(simplifyAndGroup).join("|"));
