+++ src/somethingBut.ts
@@ -1,4 +1,4 @@
-import { simpleExp } from "./utils";
+import simpleExp from "./util/simple-exp";
 
 const somethingBut = simpleExp<string>(exp => `[^${exp}]+`);
 export default somethingBut;
