--- src/types.ts
@@ -1,1 +1,0 @@
-export type Expression = RegExp | string | number;
