+++ src/types/expression.ts
@@ -1,0 +1,3 @@
+type Expression = RegExp | string | number;
+
+export default Expression;
