+++ src/util/simple-exp.ts
@@ -1,0 +1,8 @@
+import simplifyAndGroup from "./simplify-and-group";
+import Expression from "../types/expression";
+
+export default function simpleExp<ExpType extends Expression = Expression>(
+  callback: (exp: string) => string
+) {
+  return (input: ExpType) => new RegExp(callback(simplifyAndGroup(input)));
+}
