+++ src/util/simplify-and-group.ts
@@ -1,0 +1,6 @@
+import sanitize from "./sanitize";
+import Expression from "../types/expression";
+
+export default function simplifyAndGroup(expression: Expression) {
+  return `(?:${sanitize(expression)})`;
+}
